test retaining history with 'git mv' and changes
